requirements.txt
appdirs==1.4.4
asgiref==3.7.1
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
Django==4.2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
fonttools==4.39.4
freezegun==1.2.2
frozendict==2.3.8
html5lib==1.1
idna==3.4
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
multitasking==0.0.11
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
Quantsbin==1.0.3
requests==2.31.0
scipy==1.10.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
webencodings==0.5.1
yfinance==0.2.18
